pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emailgenius"
version = "0.1.0"
description = "MVP B2B web discovery + Transizione 5.0 scoring + outreach draft"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "EmailGenius"}
]
dependencies = [
  "playwright>=1.54.0",
  "PyYAML>=6.0.1",
  "psycopg[binary]>=3.2.1",
  "openai>=1.59.0",
  "gspread>=6.1.2",
  "google-auth>=2.35.0",
  "pypdf>=5.0.1",
  "python-docx>=1.1.2"
]

[project.scripts]
emailgenius = "emailgenius.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
